CoreInterface and initializeWorkerMain are now part of the features object
