+++ src/experimental/features/__tests__/local.test.ts
@@ -1,5 +1,7 @@
 import { describe, it, expect } from "vitest";
+import initializeWorkerMain from "../../../core/main/worker";
 import type { IFeaturesObject } from "../../../features/types";
+import { MonoThreadCoreInterface } from "../../../main_thread/core_interface/monothread";
 import MultiThreadContentInitializer from "../../../main_thread/init/multi_thread_content_initializer";
 import local from "../../../transports/local";
 import addLocalManifestFeature from "../local";
@@ -10,9 +12,16 @@ describe("Features list - LOCAL_MANIFEST", () => {
     addLocalManifestFeature(featureObject);
     expect(featureObject).toEqual({
       transports: { local },
-      mainThreadMediaSourceInit: MultiThreadContentInitializer,
+      monothread: {
+        init: MultiThreadContentInitializer,
+        coreInterface: MonoThreadCoreInterface,
+        workerMain: initializeWorkerMain,
+      },
     });
     expect(featureObject.transports.local).toBe(local);
-    expect(featureObject.mainThreadMediaSourceInit).toBe(MultiThreadContentInitializer);
+    expect(featureObject.monothread).not.toBe(null);
+    expect(featureObject.monothread?.init).toBe(MultiThreadContentInitializer);
+    expect(featureObject.monothread?.coreInterface).toBe(MonoThreadCoreInterface);
+    expect(featureObject.monothread?.workerMain).toBe(initializeWorkerMain);
   });
 });
