+++ src/experimental/features/multi_thread.ts
@@ -1,4 +1,5 @@
 import type { IFeaturesObject } from "../../features/types";
+import { WorkerCoreInterface } from "../../main_thread/core_interface/multithread";
 import MultiThreadContentInitializer from "../../main_thread/init/multi_thread_content_initializer";
 
 /**
@@ -6,7 +7,10 @@ import MultiThreadContentInitializer from "../../main_thread/init/multi_thread_c
  * @param {Object} features
  */
 function addMultiThreadFeature(features: IFeaturesObject): void {
-  features.multithread = { init: MultiThreadContentInitializer };
+  features.multithread = {
+    init: MultiThreadContentInitializer,
+    coreInterface: WorkerCoreInterface,
+  };
 }
 
 export { addMultiThreadFeature as MULTI_THREAD };
