+++ src/features/list/media_source_main.ts
@@ -1,3 +1,5 @@
+import initializeWorkerMain from "../../core/main/worker";
+import { MonoThreadCoreInterface } from "../../main_thread/core_interface/monothread";
 import MultiThreadContentInitializer from "../../main_thread/init/multi_thread_content_initializer";
 import type { IFeaturesObject } from "../types";
 
@@ -6,7 +8,11 @@ import type { IFeaturesObject } from "../types";
  * @param {Object} features
  */
 function addMediaSourceMainFeature(features: IFeaturesObject): void {
-  features.mainThreadMediaSourceInit = MultiThreadContentInitializer;
+  features.monothread = {
+    init: MultiThreadContentInitializer,
+    coreInterface: MonoThreadCoreInterface,
+    workerMain: initializeWorkerMain,
+  };
 }
 
 export { addMediaSourceMainFeature as MEDIA_SOURCE_MAIN };
