+++ src/main_thread/api/public_api.ts
@@ -32,7 +32,6 @@ import getStartDate from "../../compat/get_start_date";
 import hasMseInWorker from "../../compat/has_mse_in_worker";
 import hasWorkerApi from "../../compat/has_worker_api";
 import config from "../../config";
-import initializeWorkerMain from "../../core/main/worker";
 import type { ISegmentSinkMetrics } from "../../core/segment_sinks/segment_sinks_store";
 import type {
   IAdaptationChoice,
@@ -40,7 +39,6 @@ import type {
   IABRThrottlers,
   IBufferType,
 } from "../../core/types";
-import { MonoThreadCoreInterface, WorkerCoreInterface } from "../../core_interface";
 import type { IDefaultConfig } from "../../default_config";
 import type { IErrorCode, IErrorType } from "../../errors";
 import { ErrorCodes, ErrorTypes, formatError, MediaError } from "../../errors";
@@ -1058,16 +1056,16 @@ class Player extends EventEmitter<IPublicAPIEvent> {
         ) &&
         typeof options.representationFilter !== "function";
       if (mode === "main" || (mode === "auto" && !canRunInMultiThread)) {
-        if (features.mainThreadMediaSourceInit === null) {
+        if (features.monothread === null) {
           throw new Error(
             "Cannot load video, neither in a WebWorker nor with the " +
               "`MEDIA_SOURCE_MAIN` feature",
           );
         }
         log.info("API", "Initializing MediaSource mode in the main thread");
-        const coreInterface = new MonoThreadCoreInterface();
+        const coreInterface = new features.monothread.coreInterface();
         const coreInterfaceCallbacks = coreInterface.getCallbacks();
-        initializeWorkerMain(
+        features.monothread.workerMain(
           coreInterfaceCallbacks.setCoreMessageReceiver,
           coreInterfaceCallbacks.sendCoreMessage,
         );
@@ -1083,7 +1081,7 @@ class Player extends EventEmitter<IPublicAPIEvent> {
             timestamp: getMonotonicTimeStamp(),
           },
         });
-        initializer = new features.mainThreadMediaSourceInit({
+        initializer = new features.monothread.init({
           coreInterface,
           adaptiveOptions,
           autoPlay,
@@ -1119,7 +1117,7 @@ class Player extends EventEmitter<IPublicAPIEvent> {
         useWorker = true;
         log.info("API", "Initializing MediaSource mode in a WebWorker");
         initializer = new features.multithread.init({
-          coreInterface: new WorkerCoreInterface(this._priv_worker),
+          coreInterface: new features.multithread.coreInterface(this._priv_worker),
           adaptiveOptions,
           autoPlay,
           bufferOptions,
