+++ src/main_thread/core_interface/types.ts
@@ -1,0 +1,5 @@
+import type CoreInterface from "./base";
+import type { MonoThreadCoreInterface } from "./monothread";
+import type { WorkerCoreInterface } from "./multithread";
+export type { MonoThreadCoreInterface, WorkerCoreInterface };
+export default CoreInterface;
